+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 Notable changes to this project.
 
 ## [dev]
+- Remove submission `version` parameter (PR #75)
+
+## [2.11.0] - 2022-03-29
 - add tc, fnc, fncV3 to get_leaderboard
 - add icRank, icRep to SignalsAPI.get_leaderboard
 - add tcRank, tcRep, fncV3Rep and fncV3Rank to daily_model_performances